Fix goss tests and add webhook trigger
We want the docker image to be built only when tests pass.
I'm going to try triggering the Docker Hub build via a webhook. It'll build the image
twice (once on Github, once on Docker Hub) but I'm not sure publishing from Github
actions would work. I'll try it though.