Split Frontend and PHP builders
PHP Builder needs composer and php extensions.
Frontend builder needs nodejs and yarn in order to run webpack.
Splitting the two allows to separate concerns and use specialized images.

Building the frontend on alpine was a real pain in the *** because of all the
dev dependencies of koel. pngquant can't install its binary, tries to compile
itself but fails without 240MB of dependencies, which kills the whole purpose
of building on alpine.

How to test:
- docker build -t koeldev:latest .
- GOSS_SLEEP=0.5 dgoss run koeldev:latest