Add a docker-compose file for devs
Having a docker-compose file specifically for devs lets us rebuild the
image and test it with a real database and library.

How to test:
Edit the .env.dev file (which is ignored to avoid leaking passwords).
Fill the following env variables inside:

  MYSQL_ROOT_PASSWORD=<your root password>
  MYSQL_PASSWORD=<your koel DB password>
  DB_PASSWORD=<same koel DB password>

Then:
  docker-compose -f docker-compose.dev.yml up -d
Log in the web container:
  docker exec -it koeldev bash
Init koel in the web container (it will fail when it tries to run yarn)
  php artisan koel:init
Then log in with the credentials you provided at koel:init.