Do not rebuild php extensions that are already loaded
The build showed warnings because xml, curl and mbstring are already
loaded by default. No need to rebuild them.
libpng-dev was also included in the php-build image, but is only needed
for frontend stuff, so it was removed.

Also I added a volume for the .env file for the dev compose. Koel will
not let you login again if the JWT_SECRET or the APP_KEY change, so they
need to be persisted (or you need to reinit everytime).

How to test:
Build and start Koel & its database:
  docker-compose -f docker-compose.dev.yml up -d --build
In the web container (it runs until "frontend stuff", but it's okay)
  php artisan koel:init

You should have a .env.koel file with some variables.
It is ignored in git to avoid sharing secret keys.