Use alpine for php-builder
We do not need apache or debian to run 'git clone' and 'composer
install'. With fewer layers, it might be faster to build.
Interestingly, it looks like changing the PHP / composer version changed
the build. The post-install scripts now work and the .env.example in
koel's repo is copied to .env.

This means I've had to adapt the entrypoint to detect when there is
already an APP_KEY env variable or when the .env file already contains a
key.