Target multiple architectures (follow-up to #28) (#31)
* Revert and fix commit 3198879

Commit 3198879 prepended the CYPRESS_INSTALL_BINARY=0 variable to
the install command in the ressources/assets directory, however
this component does not require cypress.

The root component however does. This environment variable should
be prepended to the second install command.

* Add 'install buildx' step

This is just blindly following this blogpost:
https://www.docker.com/blog/multi-arch-build-and-images-the-simple-way/

* Target multiple architectures

See this blogpost for more on how this works:
https://www.docker.com/blog/multi-arch-build-and-images-the-simple-way/

* Add python2 dependency to front-builder

This error message shows it:
https://github.com/Filius-Patris/docker-koel/runs/915997770?check_suite_focus=true#step:5:8914

This issue comment suggests why:
https://github.com/sass/node-sass/issues/1176#issuecomment-323035618
TL;DR: Building some component requires python. This would normally be
precompiled and fetched, but arm doesn't have precompiled packages.

* Increase network timeout

The CICD errors seems to be tied to this:
https://github.com/yarnpkg/yarn/issues/4890#issuecomment-348479446

* Try building the 3 architectures at the end of the CI.

This is meant to test the CI. If it works, it will be done in
publish.yml instead of dockerimage.yml.

* Update CI/CD pipelines

For more details, see the discussions on PR #31.

With this commit, some targets in the Makefile are unneeded, but
I left them anyway.

* Correcting the tags: argument

I wasn't understanding the docs correctly.

* Refine CI/CD pipelines more

With this commit the pipeline completely cut reliance to the Makefile.
This helps to make everything more obvious by only looking at the pipeline
yaml file.

Co-authored-by: Joris MASSON <hyzual@gmail.com>