Improve the Makefile
While trying to get the repo up and running again on a new machine, I
struggled a bit with the Makefile. It now shows the help target by
default. It also creates .env.koel and .env.dev files first, otherwise
docker-compose is not happy and does not let you run "make start".

.env.koel contains koel's env file, with things like JWT_SECRET.
.env.dev contains passwords for MySQL such as the password for the root
user and for the "koel" user.